# Rearrange struct fields to save memory