--- a/comp/core/workloadmeta/def/params.go
+++ b/comp/core/workloadmeta/def/params.go
@@ -7,8 +7,8 @@ package workloadmeta
 
 // Params provides the kind of agent we're instantiating workloadmeta for
 type Params struct {
-	AgentType  AgentType
 	InitHelper InitHelper
+	AgentType  AgentType
 }
 
 // NewParams creates a Params struct with the default NodeAgent configuration