<!--
* Contributors are encouraged to read our [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Please fill the below sections if possible with relevant information or links.
-->
### What does this PR do?

### Motivation
Run fieldalignment

```
# [github.com/DataDog/datadog-agent/comp/core/workloadmeta/def]
comp/core/workloadmeta/def/params.go:9:13: struct with 16 pointer bytes could be 8
comp/core/workloadmeta/def/types.go:237:17: struct with 48 pointer bytes could be 40
comp/core/workloadmeta/def/types.go:317:21: struct with 120 pointer bytes could be 104
comp/core/workloadmeta/def/types.go:347:20: struct with 32 pointer bytes could be 24
comp/core/workloadmeta/def/types.go:425:25: struct with 72 pointer bytes could be 56
comp/core/workloadmeta/def/types.go:483:19: struct with 168 pointer bytes could be 160
comp/core/workloadmeta/def/types.go:531:16: struct with 560 pointer bytes could be 544
comp/core/workloadmeta/def/types.go:661:31: struct with 24 pointer bytes could be 16
comp/core/workloadmeta/def/types.go:695:20: struct with 360 pointer bytes could be 336
comp/core/workloadmeta/def/types.go:826:25: struct with 88 pointer bytes could be 80
comp/core/workloadmeta/def/types.go:873:27: struct with 144 pointer bytes could be 136
comp/core/workloadmeta/def/types.go:969:14: struct with 312 pointer bytes could be 304
comp/core/workloadmeta/def/types.go:1061:29: struct with 248 pointer bytes could be 224
comp/core/workloadmeta/def/types.go:1077:26: struct with 72 pointer bytes could be 48
comp/core/workloadmeta/def/types.go:1086:11: struct with 64 pointer bytes could be 56
comp/core/workloadmeta/def/types.go:1189:14: struct with 88 pointer bytes could be 72
comp/core/workloadmeta/def/types.go:1278:21: struct with 40 pointer bytes could be 24
comp/core/workloadmeta/def/types.go:1285:12: struct with 24 pointer bytes could be 16
comp/core/workloadmeta/def/types.go:1334:18: struct with 32 pointer bytes could be 16
comp/core/workloadmeta/def/filter_test.go:18:13: struct with 104 pointer bytes could be 96
comp/core/workloadmeta/def/filter_test.go:63:13: struct with 32 pointer bytes could be 24
comp/core/workloadmeta/def/filter_test.go:160:13: struct with 24 pointer bytes could be 16
comp/core/workloadmeta/def/filter_test.go:256:13: struct with 40 pointer bytes could be 32
```

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes. If you validated changes manually and not through automated tests, add context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask reviewers to do it. Describe how to set up an environment for manual tests in the PR description. Manual validation is expected to happen on every commit before merge.

Any manual validation step should then map to an automated test. Manual validation should not substitute automation, minus exceptions not supported by test tooling yet.
-->
Create a benchmark [test](https://github.com/DataDog/datadog-agent/pull/32589/files#diff-afaa3674002bea7865da86b3db5c940ee980aea2aa1e9fc5b3ab215d62b7fa82R72)

Take collector event for example: 16 bytes saved
`comp/core/workloadmeta/def/types.go:1278:21: struct with 40 pointer bytes could be 24
`
go test -bench=. -benchtime= 200000x
Old
```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/datadog-agent/comp/core/workloadmeta/def
cpu: Apple M1 Max
BenchmarkProcessCollectorEvent/Run_collector_event_bench_test_on_container-10            200000             18007 ns/op
BenchmarkProcessCollectorEvent/Run_collector_event_bench_test_on_kubernetes-10           200000             17095 ns/op
BenchmarkProcessCollectorEvent/Run_collector_event_bench_test_on_processes-10            200000             16513 ns/op
BenchmarkProcessEntity/Run_process_entity_bench_test_on_container-10                      200000             47944 ns/op
BenchmarkProcessEntity/Run_process_entity_bench_test_on_-10                               200000             47806 ns/op
BenchmarkProcessEntity/Run_process_entity_bench_test_on_process-10                        200000              6810 ns/op
PASS
ok      github.com/DataDog/datadog-agent/comp/core/workloadmeta/def     31.394s
```

New
```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/datadog-agent/comp/core/workloadmeta/def
cpu: Apple M1 Max
BenchmarkProcessCollectorEvent/Run_collector_event_bench_test_on_container-10            200000             12384 ns/op
BenchmarkProcessCollectorEvent/Run_collector_event_bench_test_on_kubernetes-10           200000             10668 ns/op
BenchmarkProcessCollectorEvent/Run_collector_event_bench_test_on_processes-10            200000             10491 ns/op
BenchmarkProcessEntity/Run_process_entity_bench_test_on_container-10                      200000             40858 ns/op
BenchmarkProcessEntity/Run_process_entity_bench_test_on_-10                               200000             41816 ns/op
BenchmarkProcessEntity/Run_process_entity_bench_test_on_process-10                        200000              6147 ns/op
PASS
ok      github.com/DataDog/datadog-agent/comp/core/workloadmeta/def     25.037s
```

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->